bot/requirements.txt
aiogram
python-dotenv
aiohttp
openai
